fix(core): merge duplicate LithicConfig definitions
Two LithicConfig classes existed: a pydantic BaseSettings in config.py and a
dataclass in config_validation.py. The package __init__.py imported from both
in succession, so the dataclass silently shadowed the pydantic version --
losing env_prefix=LITHIC_ binding and the asa_* fields (a real bug).

Merge into one canonical pydantic class in config.py, add base_url and a
validate() method to subsume the dataclass surface, and have config_validation
re-export from config so SardisConfig keeps working with field(default_factory).

Ref: ~/project-directions/sardis-python-sdk-redesign.md, Implementation Risks #4.
