+++ packages/sardis-core/src/sardis_v2_core/__init__.py
@@ -79,7 +79,6 @@ from .config_validation import (
     DatabaseConfig,
     EllipticConfig,
     Environment,
-    LithicConfig,
     PersonaConfig,
     SardisConfig,
     load_config_from_env,
@@ -629,7 +628,6 @@ __all__ = [
     "TurnkeyConfigModel",
     "PersonaConfig",
     "EllipticConfig",
-    "LithicConfig",
     "ChainConfigModel",
     "APIServerConfig",
     "Environment",
